consensus-state-store: rename directory cs -> consensus_state
